# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAISON Luxury Fashion E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /search:
    get:
      operationId: searchProducts
      tags: [search]
      summary: Search products
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /checkout/create-payment-intent:
    post:
      operationId: createPaymentIntent
      tags: [orders]
      summary: Create Stripe payment intent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentRequest"
      responses:
        "200":
          description: Payment intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResponse"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [orders]
      summary: Validate coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponRequest"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResponse"
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        sku:
          type: string
        stock:
          type: integer
        categoryId:
          type: string
        categoryName:
          type: string
        categorySlug:
          type: string
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isArchived:
          type: boolean
        averageRating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - price
        - images
...
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
...
    CreatePaymentIntentRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        cartId:
          type: string
        guestId:
          type: string
      required:
        - amount
    PaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
      required:
        - clientSecret
        - paymentIntentId
    ValidateCouponRequest:
      type: object
      properties:
        code:
          type: string
        orderTotal:
          type: number
      required:
        - code
        - orderTotal
    CouponValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        discount:
          type: number
        type:
          type: string
          enum: [PERCENT, FIXED]
        message:
          type: string
      required:
        - valid